+++ tests/erlang_tests/test_send.erl
@@ -20,15 +20,74 @@
 
 -module(test_send).
 
--export([start/0, send2/2]).
+-export([start/0]).
 
 start() ->
-    send2(5, 1) + send2(self(), -1).
+    Dead = spawn(fun() -> ok end),
+    Sent = send(self(), 32),
+    receive
+        Sent ->
+            Recv = Sent;
+        _Any ->
+            Recv = 64
+    after 5000 ->
+        Recv = 128
+    end,
+    T1 = Sent - Recv,
+    T2 = send_mal(5, 3),
+    T3 = send_bad_atom(bogus, 4),
+    T4 = send_dead(Dead, 6),
+    T5 = send_registered(8),
+    T1 + T2 + T3 + T4 + T5.
 
-send2(A, B) ->
+send(A, B) ->
     try erlang:send(A, B) of
-        B -> -1;
-        Any -> Any
+        B -> B;
+        _Any -> -1
+    catch
+        error:badarg -> B - 2;
+        _:_ -> -4
+    end.
+
+send_mal(A, B) ->
+    try erlang:send(A, B) of
+        B -> B;
+        _Any -> -1
+    catch
+        error:badarg -> B - 3;
+        _:_ -> -4
+    end.
+
+send_bad_atom(A, B) ->
+    try erlang:send(A, B) of
+        B -> B;
+        _Any -> -1
+    catch
+        error:badarg -> B - 4;
+        _:_ -> -4
+    end.
+
+send_dead(A, B) ->
+    try erlang:send(A, B) of
+        B -> B - 6;
+        _Any -> -1
+    catch
+        error:badarg -> -2;
+        _:_ -> -4
+    end.
+
+send_registered(B) ->
+    erlang:register(listen, self()),
+    try erlang:send(listen, B) of
+        B ->
+            receive
+                B -> B - 8;
+                _Any -> -1
+            after 5000 ->
+                512
+            end;
+        _Any ->
+            -1
     catch
         error:badarg -> -2;
         _:_ -> -4